requirements.txt
certifi
python-dotenv